+++ embodisteer/guidance/__init__.py
@@ -3,7 +3,6 @@
 from .cbf import solve_batched_cbf_qp
 from .diffusion import (
     flatten_obstacle_info,
-    get_guidance_strength,
     get_pred_x0,
     rel_action_obstacle_loss,
 )
@@ -11,7 +10,6 @@ from .schedule import guidance_scale_at, logistic_guidance_strength
 
 __all__ = [
     "flatten_obstacle_info",
-    "get_guidance_strength",
     "get_pred_x0",
     "guidance_scale_at",
     "logistic_guidance_strength",